added readme to example
