+++ examples/websocket/Readme.md
@@ -1,0 +1,3 @@
+# WebSocket transport binding
+
+Run the example with `cargo run --bin sender` or `cargo run --bin receiver`
